requirements.txt
django
channels
daphne
redis
channels-redis
celery[redis]
django-celery-results
django-celery-beat
djangorestframework
djangorestframework-gis
django-filter
drf-spectacular
requests
pandas
geopandas
gtfs-realtime-bindings
psycopg2-binary
gunicorn
mkdocs-material
python-decouple
